Refactor/small changes (#129)
* Remove unnecessary isSpace function

* Refactor BrowseView and CollectionPictureDisplay to one component

* Refactor CollectionPictureDisplay tests to BrowseView tests

Thereby simplify them and adjust associated mocks
